requirements.txt
openai-whisper          # state-of-the-art, fully local STT model
requests
tqdm
requests